#120 Move defaultNamespace and mergeNamespace to separate files
