+++ packages/namespaces/src/attachNamespace.js
@@ -1,31 +1,4 @@
-import { curry, mergeDeepRight, mergeDeepLeft } from 'ramda';
-import { isFunction } from 'ramda-extension';
-
-import getNamespaceByAction from './getNamespaceByAction';
-
-const mergeNamespace = curry((isForced, namespace, action) => {
-	if (!namespace) {
-		return action;
-	}
-
-	if (isFunction(action)) {
-		const nextAction = (...args) => action(...args);
-
-		if (isForced) {
-			nextAction.meta = { namespace };
-		} else {
-			nextAction.meta = { namespace: getNamespaceByAction(action) || namespace };
-		}
-
-		return nextAction;
-	}
-
-	if (isForced) {
-		return mergeDeepLeft({ meta: { namespace } }, action);
-	}
-
-	return mergeDeepRight({ meta: { namespace } }, action);
-});
+import mergeNamespace from './mergeNamespace';
 
 /**
  * Associates an action with a namespace, overwriting any previous namespace.
@@ -33,12 +6,6 @@ const mergeNamespace = curry((isForced, namespace, action) => {
  * @param {?string} namespace namespace to attach
  * @param {Object} action action to add the namespace to
  */
-export const attachNamespace = mergeNamespace(true);
+const attachNamespace = mergeNamespace(true);
 
-/**
- * Associates an action with a namespace unless it is already associated with some namespace.
- *
- * @param {?string} namespace namespace to attach
- * @param {Object} action action to add the namespace to
- */
-export const defaultNamespace = mergeNamespace(false);
+export default attachNamespace;
