+++ packages/namespaces/src/index.js
@@ -1,4 +1,5 @@
-export { attachNamespace, defaultNamespace } from './attachNamespace';
+export { default as attachNamespace } from './attachNamespace';
+export { default as defaultNamespace } from './defaultNamespace';
 export { default as getNamespaceByAction } from './getNamespaceByAction';
 export { default as isActionFromNamespace } from './isActionFromNamespace';
 export * from './constants';
